Rename snp to get_report_types
