+++ src/guest/attestation/get_report_types.rs
